Ui Render Changes.
- `font_size` now f32, was u16 (necessary for `Canvas::World`)
- Now ui nodes doesnt stretch in `Canvas::Camera` mode and save original aspect.

+++ crates/runa_core/src/ui/handle.rs
@@ -160,7 +160,7 @@ impl TextHandle {
         }
     }
 
-    pub fn set_font_size(&self, ui: &mut UiRenderer, s: u16) {
+    pub fn set_font_size(&self, ui: &mut UiRenderer, s: f32) {
         if let Some(n) = ui.node_mut(self.0) {
             if let UiNodeKind::Text(p) = &mut n.kind {
                 p.font_size = s;
